refactor loading/writing files to `file_handler` module
