+++ leftwm/src/bin/leftwm-check.rs
@@ -1,18 +1,12 @@
 use anyhow::{bail, Result};
 use clap::{arg, command};
-use leftwm::{Config, ThemeSetting};
-use ron::{
-    extensions::Extensions,
-    ser::{to_string_pretty, PrettyConfig},
-    Options,
+use leftwm::utils::file_handler::{load_from_file, write_to_file};
+use leftwm::ThemeSetting;
+use std::{
+    env, fs,
+    os::unix::fs::PermissionsExt,
+    path::{Path, PathBuf},
 };
-use std::env;
-use std::fs;
-use std::fs::File;
-use std::io::Write;
-use std::os::unix::fs::PermissionsExt;
-use std::path::Path;
-use std::path::PathBuf;
 use xdg::BaseDirectories;
 
 #[tokio::main]
@@ -51,6 +45,53 @@ async fn main() -> Result<()> {
         return Ok(());
     }
 
+    pub(crate) fn check_elogind(verbose: bool) -> Result<()> {
+        // We assume that if it is in the path it's all good
+        // We also cross-reference the ENV variable
+        match (
+            std::env::var("XDG_RUNTIME_DIR"),
+            leftwm::is_program_in_path("loginctl"),
+        ) {
+            (Ok(val), true) => {
+                if verbose {
+                    println!(":: XDG_RUNTIME_DIR: {val}, LOGINCTL OKAY");
+                }
+
+                println!("\x1b[0;92m    -> Environment OK \x1b[0m");
+
+                Ok(())
+            }
+            (Ok(val), false) => {
+                if verbose {
+                    println!(":: XDG_RUNTIME_DIR: {val}, LOGINCTL not installed");
+                }
+
+                println!("\x1b[0;92m    -> Environment OK (has XDG_RUNTIME_DIR) \x1b[0m");
+
+                Ok(())
+            }
+            (Err(e), false) => {
+                if verbose {
+                    println!(":: XDG_RUNTIME_DIR_ERROR: {e:?}, LOGINCTL BAD");
+                }
+
+                bail!(
+                    "Elogind not installed/operating and no alternative XDG_RUNTIME_DIR is set. \
+                See https://github.com/leftwm/leftwm/wiki/XDG_RUNTIME_DIR for more information."
+                );
+            }
+            (Err(e), true) => {
+                if verbose {
+                    println!(":: XDG_RUNTIME_DIR: {e:?}, LOGINCTL OKAY");
+                }
+                println!(
+                    "\x1b[1;93mWARN: Elogind/systemd installed but XDG_RUNTIME_DIR not set.\nThis may be because elogind isn't started. \x1b[0m",
+                );
+                Ok(())
+            }
+        }
+    }
+
     match check_enabled_features() {
         Ok(_) => {}
         Err(err) => {
@@ -81,121 +122,6 @@ async fn main() -> Result<()> {
     check_theme(verbose);
 
     Ok(())
-}
-
-/// Loads configuration from either specified file (preferred) or default.
-/// # Errors
-///
-/// Errors if file cannot be read. Indicates filesystem error
-/// (inadequate permissions, disk full, etc.)
-/// If a path is specified and does not exist, returns `LeftError`.
-pub fn load_from_file(fspath: Option<&str>, verbose: bool) -> Result<Config> {
-    let config_filename = if let Some(fspath) = fspath {
-        println!("\x1b[1;35mNote: Using file {fspath} \x1b[0m");
-        PathBuf::from(fspath)
-    } else {
-        let ron_file = BaseDirectories::with_prefix("leftwm")?.place_config_file("config.ron")?;
-        let toml_file = BaseDirectories::with_prefix("leftwm")?.place_config_file("config.toml")?;
-        if Path::new(&ron_file).exists() {
-            ron_file
-        } else if Path::new(&toml_file).exists() {
-            println!(
-                "\x1b[1;93mWARN: TOML as config format is about to be deprecated.
-      Please consider migrating to RON manually or by using `leftwm-check -m`.\x1b[0m"
-            );
-            toml_file
-        } else {
-            let config = Config::default();
-            write_to_file(&ron_file, &config)?;
-            return Ok(config);
-        }
-    };
-
-    if verbose {
-        dbg!(&config_filename);
-    }
-    let contents = fs::read_to_string(&config_filename)?;
-    if verbose {
-        dbg!(&contents);
-    }
-    if config_filename.as_path().extension() == Some(std::ffi::OsStr::new("ron")) {
-        let ron = Options::default().with_default_extension(Extensions::IMPLICIT_SOME);
-        let config: Config = ron.from_str(&contents)?;
-        Ok(config)
-    } else {
-        let config = toml::from_str(&contents)?;
-        Ok(config)
-    }
-}
-
-fn write_to_file(ron_file: &Path, config: &Config) -> Result<(), anyhow::Error> {
-    let ron_pretty_conf = PrettyConfig::new()
-        .depth_limit(2)
-        .extensions(Extensions::IMPLICIT_SOME);
-    let ron = to_string_pretty(&config, ron_pretty_conf)?;
-    let comment_header = String::from(
-        r#"//  _        ___                                      ___ _
-// | |      / __)_                                   / __|_)
-// | | ____| |__| |_ _ _ _ ____      ____ ___  ____ | |__ _  ____    ____ ___  ____
-// | |/ _  )  __)  _) | | |    \    / ___) _ \|  _ \|  __) |/ _  |  / ___) _ \|  _ \
-// | ( (/ /| |  | |_| | | | | | |  ( (__| |_| | | | | |  | ( ( | |_| |  | |_| | | | |
-// |_|\____)_|   \___)____|_|_|_|   \____)___/|_| |_|_|  |_|\_|| (_)_|   \___/|_| |_|
-// A WindowManager for Adventurers                         (____/
-// For info about configuration please visit https://github.com/leftwm/leftwm/wiki
-
-"#,
-    );
-    let ron_with_header = comment_header + &ron;
-    let mut file = File::create(ron_file)?;
-    file.write_all(ron_with_header.as_bytes())?;
-    Ok(())
-}
-
-fn check_elogind(verbose: bool) -> Result<()> {
-    // We assume that if it is in the path it's all good
-    // We also cross-reference the ENV variable
-    match (
-        std::env::var("XDG_RUNTIME_DIR"),
-        leftwm::is_program_in_path("loginctl"),
-    ) {
-        (Ok(val), true) => {
-            if verbose {
-                println!(":: XDG_RUNTIME_DIR: {val}, LOGINCTL OKAY");
-            }
-
-            println!("\x1b[0;92m    -> Environment OK \x1b[0m");
-
-            Ok(())
-        }
-        (Ok(val), false) => {
-            if verbose {
-                println!(":: XDG_RUNTIME_DIR: {val}, LOGINCTL not installed");
-            }
-
-            println!("\x1b[0;92m    -> Environment OK (has XDG_RUNTIME_DIR) \x1b[0m");
-
-            Ok(())
-        }
-        (Err(e), false) => {
-            if verbose {
-                println!(":: XDG_RUNTIME_DIR_ERROR: {e:?}, LOGINCTL BAD");
-            }
-
-            bail!(
-                "Elogind not installed/operating and no alternative XDG_RUNTIME_DIR is set. \
-                See https://github.com/leftwm/leftwm/wiki/XDG_RUNTIME_DIR for more information."
-            );
-        }
-        (Err(e), true) => {
-            if verbose {
-                println!(":: XDG_RUNTIME_DIR: {e:?}, LOGINCTL OKAY");
-            }
-            println!(
-                "\x1b[1;93mWARN: Elogind/systemd installed but XDG_RUNTIME_DIR not set.\nThis may be because elogind isn't started. \x1b[0m",
-            );
-            Ok(())
-        }
-    }
 }
 
 /// Checks if `.config/leftwm/theme/current/` is a valid path
