+++ leftwm/src/lib.rs
@@ -7,3 +7,4 @@ mod theme_setting;
 pub use command::*;
 pub use config::*;
 pub use theme_setting::*;
+// pub use utils::file_handler::*;
