+++ leftwm/src/theme_setting.rs
@@ -1,8 +1,7 @@
+use crate::utils::file_handler::load_theme_file;
 use anyhow::Result;
 use leftwm_core::models::{Gutter, Margins};
-use ron::{extensions::Extensions, Options};
 use serde::{Deserialize, Serialize};
-use std::fs;
 use std::path::Path;
 
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
@@ -50,18 +49,6 @@ impl Default for ThemeSetting {
             background_color: Some("#333333".to_owned()),
             on_new_window_cmd: None,
         }
-    }
-}
-
-fn load_theme_file(path: impl AsRef<Path>) -> Result<ThemeSetting> {
-    let contents = fs::read_to_string(&path)?;
-    if path.as_ref().extension() == Some(std::ffi::OsStr::new("ron")) {
-        let ron = Options::default().with_default_extension(Extensions::IMPLICIT_SOME);
-        let from_file: ThemeSetting = ron.from_str(&contents)?;
-        Ok(from_file)
-    } else {
-        let from_file: ThemeSetting = toml::from_str(&contents)?;
-        Ok(from_file)
     }
 }
